% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ashplots.R
\name{ash_points}
\alias{ash_points}
\title{Compute knot points of an average shifted histogram}
\usage{
ash_points(x, binwidth = NULL, adjust = 1)
}
\arguments{
\item{x}{A numeric vector}

\item{binwidth}{The width of the histogram bins.  If \code{NULL} (the default) the
binwidth will be chosen so that approximately 10 bins cover the data.  \code{adjust}
can be used to to increase or decrease \code{binwidth}.}

\item{adjust}{A number used to scale \code{binwidth}.}
}
\value{
A data frame containing x and y coordinates of the resulting ASH plot.
}
\description{
Mainly a utility for the \pkg{lattice} and \pkg{ggplot2} plotting
functions, \code{ash_points()} returns the points to be plotted.
}
