% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{compare}
\alias{compare}
\title{Campare two numeric vectors}
\usage{
compare(x, y, verbose = FALSE)
}
\arguments{
\item{x, y}{numeric verctors to be compared item by item}

\item{verbose}{a logical indicating whether verbose labeling is desired}
}
\description{
This wrapper around \code{sign()} provides a more intuitive labeling.
}
\examples{
tally( ~ compare(mcs, pcs), data = HELPrct)
tally( ~ compare(mcs, pcs, verbose = TRUE), data = HELPrct)
tally( ~ compare(sexrisk, drugrisk), data = HELPrct)
}
