% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer-transformation.R
\name{infer_transformation}
\alias{infer_transformation}
\title{Infer a Back-Transformation}
\usage{
infer_transformation(formula, warn = TRUE)
}
\arguments{
\item{formula}{A formula as used by, for example, \code{\link{lm}}.}

\item{warn}{A logical.}
}
\value{
A function.
}
\description{
For a handful of transformations on y, infer the reverse transformation.  If the
transformation is not recognized, return the identity function.  This is primarily
intended to be used for setting a default value in other functions.
}
