% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally.R
\name{logical2factor}
\alias{logical2factor}
\alias{logical2factor.default}
\alias{logical2factor.data.frame}
\title{Convert logical vector into factor}
\usage{
logical2factor(x, ...)

\method{logical2factor}{default}(x, ...)

\method{logical2factor}{data.frame}(x, ...)
}
\arguments{
\item{x}{a vector or data frame}

\item{\dots}{additional arguments (currently ignored)}
}
\value{
If \code{x} is a vector either \code{x} or the result
of converting \code{x} into a factor with levels \code{TRUE}
and \code{FALSE} (in that order);  if \code{x} is a data frame,
a data frame with all logicals converted to factors in this manner.
}
\description{
Turn logicals into factors with levels ordered with \code{TRUE} before \code{FALSE}.
Other inputs are returned unaltered.
}
