% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vector2df.R
\name{make_df}
\alias{make_df}
\alias{make_df.list}
\alias{make_df.matrix}
\alias{make_df.numeric}
\alias{make_df.default}
\title{Convert to a data frame}
\usage{
make_df(object, ...)

\method{make_df}{list}(object, ...)

\method{make_df}{matrix}(object, ...)

\method{make_df}{numeric}(object, ...)

\method{make_df}{default}(object, ...)
}
\arguments{
\item{object}{An object to be converted into a data frame.}

\item{...}{Additional arguments used by methods.}
}
\description{
A generic and several methods for converting objects into data frames.
}
