% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tally.R
\name{mosaic_tally}
\alias{mosaic_tally}
\alias{mosaic_count}
\title{Internal tally methods}
\usage{
mosaic_tally(x, ...)

mosaic_count(x, ...)
}
\arguments{
\item{x}{an object}

\item{...}{additional arguments passed to \code{\link[=table]{table()}}}
}
\description{
These are used to implement \code{tally()} in a way that allows {dplyr} and {mosaicCore} to co-exist.
End users should call the generics \code{tally()} and \code{count()}.
}
