% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_stats.R
\name{na.warn}
\alias{na.warn}
\title{Exclude Missing Data with Warning}
\usage{
na.warn(object, ...)
}
\arguments{
\item{object}{an \R object, typically a data frame}

\item{...}{further arguments special methods could require.}
}
\description{
Similar to \code{\link[stats:na.fail]{stats::na.exclude()}} this function excludes missing data.
When missing data are excluded, a warning message indicating the number of excluded
rows is emited as a caution for the user.
}
