% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_stats.R
\name{rhs_or_expr}
\alias{rhs_or_expr}
\title{Return rhs of a formula or expression}
\usage{
rhs_or_expr(x)
}
\arguments{
\item{x}{A formula or some other object to be quoted}
}
\description{
Return rhs of a formula or expression
}
\examples{
# This should evaluate to TRUE
rhs_or_expr(~z)
rhs_or_expr(z)
identical(rhs_or_expr(~z), rhs_or_expr(z))
}
